Tests/eval/requirements.txt
# Python evaluation test requirements
# Install with: pip install -r requirements.txt

httpx==0.27.0        # Async HTTP client for API calls
pytest==8.2.0        # Test runner
pytest-asyncio==0.23.6  # Async test support
pytest-timeout==2.3.1   # Timeout per test
pytest-cov==5.0.0    # Coverage reporting
rich==13.7.0         # Pretty terminal output for evaluate_responses.py
python-dotenv==1.0.1 # Load .env for CHATBOT_BASE_URL
